fix: durable installations, more
