+++ packages/vat-data/src/vat-data-bindings.js
@@ -1,0 +1,99 @@
+/* global globalThis */
+
+import { assert } from '@agoric/assert';
+import {
+  M,
+  makeScalarMapStore,
+  makeScalarWeakMapStore,
+  makeScalarSetStore,
+  makeScalarWeakSetStore,
+} from '@agoric/store';
+
+export {
+  M,
+  makeScalarMapStore,
+  makeScalarWeakMapStore,
+  makeScalarSetStore,
+  makeScalarWeakSetStore,
+};
+
+/** @type {import('./types').VatData} */
+let VatDataGlobal;
+if ('VatData' in globalThis) {
+  assert(globalThis.VatData, 'VatData defined in global as null or undefined');
+  VatDataGlobal = globalThis.VatData;
+} else {
+  // XXX this module has been known to get imported (transitively) in cases that
+  // never use it so we make a version that will satisfy module resolution but
+  // fail at runtime.
+  const unvailable = () => assert.fail('VatData unavailable');
+  VatDataGlobal = {
+    defineKind: unvailable,
+    defineKindMulti: unvailable,
+    defineDurableKind: unvailable,
+    defineDurableKindMulti: unvailable,
+    makeKindHandle: unvailable,
+    providePromiseWatcher: unvailable,
+    watchPromise: unvailable,
+    makeScalarBigMapStore: unvailable,
+    makeScalarBigWeakMapStore: unvailable,
+    makeScalarBigSetStore: unvailable,
+    makeScalarBigWeakSetStore: unvailable,
+  };
+}
+
+export const {
+  defineKind,
+  defineKindMulti,
+  defineDurableKind,
+  defineDurableKindMulti,
+  makeKindHandle,
+  providePromiseWatcher,
+  watchPromise,
+  makeScalarBigMapStore,
+  makeScalarBigWeakMapStore,
+  makeScalarBigSetStore,
+  makeScalarBigWeakSetStore,
+} = VatDataGlobal;
+
+/**
+ * When making a multi-facet kind, it's common to pick one facet to expose. E.g.,
+ *
+ *     const makeFoo = (a, b, c, d) => makeFooBase(a, b, c, d).self;
+ *
+ * This helper reduces the duplication:
+ *
+ *     const makeFoo = pickFacet(makeFooBase, 'self');
+ *
+ * @type {import('./types').PickFacet}
+ */
+export const pickFacet =
+  (maker, facetName) =>
+  (...args) =>
+    maker(...args)[facetName];
+// @ts-expect-error TODO statically recognize harden
+harden(pickFacet);
+
+/**
+ * Assign the values of all of the enumerable own properties from the source
+ * object to their keys in the target object.
+ *
+ * @template {{}} T
+ * @param {T} target
+ * @param {Partial<T>} source
+ */
+export const partialAssign = (target, source) => {
+  Object.assign(target, source);
+};
+// @ts-expect-error TODO statically recognize harden
+harden(partialAssign);
+
+export const provideBaggageSubtree = (baggage, name) => {
+  if (!baggage.has(name)) {
+    // TODO: maybe string keys
+    baggage.init(name, makeScalarBigMapStore(name, { durable: true }));
+  }
+  return baggage.get(name);
+};
+// @ts-expect-error TODO statically recognize harden
+harden(provideBaggageSubtree);
